+++ src/main/java/com/project/zodiac_api/dto/ClientResponseDto.java
@@ -21,12 +21,12 @@ public class ClientResponseDto {
 
     // 上升 / 天頂四軸資訊（允許 null）
     private String  ascSign;
-    private Integer ascDegreeNum;
-    private Integer ascMinuteNum;
+    private Short ascDegreeNum;
+    private Short ascMinuteNum;
 
     private String  mcSign;
-    private Integer mcDegreeNum;
-    private Integer mcMinuteNum;
+    private Short mcDegreeNum;
+    private Short mcMinuteNum;
 
     private LocalDateTime createdAt;
